move SingletonColumn into Column
