+++ yggdrasil/src/main/scala/quasar/yggdrasil/table/Column.scala
@@ -600,3 +600,95 @@ object Column {
   def isDefinedAtAll(cols: Array[Column], row: Int): Boolean =
     cols.length > 0 && cols.forall(_ isDefinedAt row)
 }
+
+abstract class SingletonColumn[A] extends ExtensibleColumn {
+
+  override def isDefinedAt(row: Int) = row == 0
+
+  override def definedAt(from: Int, to: Int): BitSet = SingletonColumn.Defined.copy
+
+}
+
+final class SingletonBoolColumn(val value: Boolean)
+    extends SingletonColumn[Boolean] with BoolColumn {
+  def apply(row: Int): Boolean = value
+}
+
+object SingletonBoolColumn {
+  lazy val SingleTrue = new SingletonBoolColumn(true)
+  lazy val SingleFalse = new SingletonBoolColumn(false)
+
+  def bitset(b: Boolean) =
+    if (b) SingletonColumn.BitSet_1
+    else SingletonColumn.BitSet_0
+
+  def apply(value: Boolean): SingletonBoolColumn =
+    if (value) SingleTrue else SingleFalse
+}
+
+final case class SingletonLongColumn(value: Long)
+    extends SingletonColumn[Long] with LongColumn {
+  def apply(row: Int): Long = value
+}
+
+final case class SingletonDoubleColumn(value: Double)
+    extends SingletonColumn[Double] with DoubleColumn {
+  def apply(row: Int): Double = value
+}
+
+final case class SingletonNumColumn(value: BigDecimal)
+    extends SingletonColumn[BigDecimal] with NumColumn {
+  def apply(row: Int): BigDecimal = value
+}
+
+final case class SingletonStrColumn(value: String)
+    extends SingletonColumn[String] with StrColumn {
+  def apply(row: Int): String = value
+}
+
+final case class SingletonOffsetDateTimeColumn(value: OffsetDateTime)
+    extends SingletonColumn[OffsetDateTime]with OffsetDateTimeColumn {
+  def apply(row: Int): OffsetDateTime = value
+}
+
+final case class SingletonOffsetTimeColumn(value: OffsetTime)
+    extends SingletonColumn[OffsetTime] with OffsetTimeColumn {
+  def apply(row: Int): OffsetTime = value
+}
+
+final case class SingletonOffsetDateColumn(value: OffsetDate)
+    extends SingletonColumn[OffsetDate] with OffsetDateColumn {
+  def apply(row: Int): OffsetDate = value
+}
+
+final case class SingletonLocalDateTimeColumn(value: LocalDateTime)
+    extends SingletonColumn[LocalDateTime] with LocalDateTimeColumn {
+  def apply(row: Int): LocalDateTime = value
+}
+
+final case class SingletonLocalTimeColumn(value: LocalTime)
+    extends SingletonColumn[LocalTime] with LocalTimeColumn {
+  def apply(row: Int): LocalTime = value
+}
+
+final case class SingletonLocalDateColumn(value: LocalDate)
+    extends SingletonColumn[LocalDate] with LocalDateColumn {
+  def apply(row: Int): LocalDate = value
+}
+
+final case class SingletonIntervalColumn(value: DateTimeInterval)
+    extends SingletonColumn[DateTimeInterval] with IntervalColumn {
+  def apply(row: Int): DateTimeInterval = value
+}
+
+object SingletonColumn {
+  val BitSet_0: BitSet = new BitSet
+
+  val BitSet_1: BitSet = {
+    val bs = new BitSet
+    bs.set(0)
+    bs
+  }
+
+  val Defined: BitSet = BitSet_1
+}
